version 1.19.7

+++ usr/local/pproxy/setup/install_wireguard.sh
@@ -18,17 +18,41 @@
 # apt-get install wireguard
 
 wg-quick down wg0
-PORT=`cat /etc/pproxy/config.ini  | grep wireport | awk '{print $3}'`
+ORPORT=`cat /etc/pproxy/config.ini  | grep wireport | tr -d ' ' | awk -F"=" '{print $2}'`
 PORT=${ORPORT:=6711}
+PUBPUB=/var/local/pproxy/wireguard-publickey
+echo "setting up wireguard on port $ORPORT"
+
+# ip_forward should be already on, just in case:
+if ! grep -q "^net.ipv4.ip_forward$" /etc/sysctl.conf;
+then
+	echo "adding ip_forward to sysctl to persist"
+	sed -i 's|#net.ipv4.ip_forward=1|net.ipv4.ip_forward=1|' /etc/sysctl.conf
+	/usr/sbin/sysctl -p
+fi
 
 configs_path=/var/local/pproxy/users/
-mkdir $configs_path
+mkdir -p $configs_path
 chown pproxy:pproxy $configs_path
 
 cd /etc/wireguard
 
 umask 077
-wg genkey | tee privatekey | wg pubkey > publickey
+if [ -e wg0.conf ];
+then
+	server_priv=`cat /etc/wireguard/wg0.conf | grep PrivateKey | awk -F" = " '{print $2}'`
+	echo $server_priv
+	if [ ! -z $server_priv ]; then
+		echo $server_priv > privatekey
+		echo $server_priv | wg pubkey > publickey
+	fi
+fi
+
+if [ ! -e privatekey ];
+then
+	wg genkey | tee privatekey | wg pubkey > publickey
+
+fi
 
 PUB_SERVER=`cat publickey`
 PRIV_SERVER=`cat privatekey`
@@ -36,8 +60,10 @@ PRIV_SERVER=`cat privatekey`
 cat > /etc/wireguard/wg0.conf << EOF
 
 [Interface]
+SaveConfig = true
 PrivateKey = $PRIV_SERVER
 Address = 10.93.76.1/24
+DNS = 1.1.1.1
 ListenPort = $PORT
 PostUp = iptables -I INPUT -p udp --dport $PORT -j ACCEPT
 PostUp = iptables -I FORWARD -i eth0 -o wg0 -j ACCEPT
@@ -59,7 +85,9 @@ wg-quick up wg0
 wg show
 systemctl enable wg-quick@wg0
 
-rm -f publickey
-rm -f privatekey
+cp publickey $PUBPUB
+chmod 0644 $PUBPUB
+# just to be safe, umask should have already set it right
+chmod 0600 privatekey
 
 cd -
